Alot of changes
First off we changed everything so that we now use the MVVM architectural pattern. After that, we removed our ProductCategory class and made one singular list for all products. We then fixed up our Products.xaml by removing our use of UserControls and instead implemented it directly into the xaml to fix our ItemsControl not working. Lastly, we added a ScrollViewer to the ItemsControl so that if we add more products, you will still be able to see all of them by using the scroll bar